refactor: move to directory
